# Conflicts:
#	.gitignore
#	src/simpleauth.authserver/Startup.cs
#	src/simpleauth.authserver/appsettings.json
#	src/simpleauth.authserver/simpleauth.authserver.csproj
#	src/simpleauth.shared/Events/Logging/ResourceOwnerAdded.cs
#	src/simpleauth/Controllers/RegistrationController.cs
#	src/simpleauth/Extensions/ServiceCollectionExtensions.cs

+++ src/simpleauth.shared/Models/GenericResult.cs
@@ -15,7 +15,7 @@
         /// The total results.
         /// </value>
         [DataMember(Name = "count")]
-        public int TotalResults { get; set; }
+        public long TotalResults { get; set; }
 
         /// <summary>
         /// Gets or sets the start index.
